+++ fltk/src/utils.rs
@@ -15,3 +15,26 @@ impl FlString for CString {
         }
     }
 }
+
+/// Convenience function to convert rgb to hex
+/// Example: 
+/// ```rust
+/// let ret = rgb2hex(0, 255, 0); println!("0x{:06x}", ret);
+/// ```
+pub fn rgb2hex(r: u8, g: u8, b: u8) -> u32 {
+    // Shouldn't fail
+    format!("0x{:02x}{:02x}{:02x}", r, g, b).parse().unwrap()
+}
+
+/// Convenience function to convert hex to rgb
+/// Example:
+/// ```rust
+/// let (r, g, b) = hex2rgb(0x000000);
+/// ```
+pub fn hex2rgb(val: u32) -> (u8, u8, u8) {
+    let hex = format!("{:06x}", val);
+    let r = u8::from_str_radix(&hex[0..2], 16).unwrap();
+    let g = u8::from_str_radix(&hex[2..4], 16).unwrap();
+    let b = u8::from_str_radix(&hex[4..6], 16).unwrap();
+    (r, g, b)
+}